# docs/openapi/backend-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/conversations/{id}/messages": {
      "get": {
        "operationId": "MessagesController_getMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_createMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
    "/conversations/{id}/read": {
      "post": {
        "operationId": "MessagesController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
...
    "/marketplace/posts/{id}/reviews": {
      "get": {
        "operationId": "MarketplaceController_getReviews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marketplace"
        ]
      },
      "post": {
        "operationId": "MarketplaceController_createReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/marketplace/products/{id}/orders": {
      "post": {
        "operationId": "MarketplaceController_createOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/marketplace/services/{id}/bookings": {
      "post": {
        "operationId": "MarketplaceController_createBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
...
    "/mentorship/posts/{id}/offers": {
      "post": {
        "operationId": "MentorshipController_createOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Mentorship"
        ]
      }
    },
    "/mentorship/relationships/{id}/status": {
      "post": {
        "operationId": "MentorshipController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Mentorship"
        ]
      }
    },
...
    "/profiles/{id}/similar": {
      "get": {
        "operationId": "ProfilesController_getSimilarProfiles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/{id}/views": {
      "post": {
        "operationId": "ProfilesController_incrementViews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/{id}": {
      "get": {
        "operationId": "ProfilesController_getProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    }
  },
  "info": {
    "title": "Vendrome API",
    "description": "Vendrome backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}